requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
chromadb==0.3.22
click==8.1.3
clickhouse-connect==0.5.23
dataclasses-json==0.5.7
docx2txt==0.8
duckdb==0.7.1
exceptiongroup==1.0.4
fastapi==0.95.1
filelock==3.12.0
Flask==2.2.3
flask-expects-json==1.7.0
frozenlist==1.3.3
fsspec==2023.5.0
greenlet==2.0.2
grpcio==1.48.1
grpcio-tools==1.48.1
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.161
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pandas==2.0.1
Pillow==9.2.0
posthog==3.0.1
protobuf==3.19.5
pydantic==1.10.7
pypdf==3.8.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.6.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.12
stability-sdk==0.2.6
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.12
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.9.2
zstandard==0.21.0
